refactor(TextInputInnerVisualSlot): moved to internal
