+++ src/TextInputWithTokens/TextInputWithTokens.tsx
@@ -9,7 +9,7 @@ import Text from '../Text'
 import {TextInputProps} from '../TextInput'
 import Token from '../Token/Token'
 import {TokenSizeKeys} from '../Token/TokenBase'
-import TextInputInnerVisualSlot from '../_TextInputInnerVisualSlot'
+import TextInputInnerVisualSlot from '../internal/components/TextInputInnerVisualSlot'
 import TextInputWrapper, {textInputHorizPadding, TextInputSizes} from '../_TextInputWrapper'
 import UnstyledTextInput from '../_UnstyledTextInput'
